feat(cli-config): Separate functionality in utility files
